@@ -12,7 +12,7 @@
     <properties>
         <org.iids.aos.install.type>agent</org.iids.aos.install.type>
         <org.iids.aos.install.path>$AGENTSCAPE_HOME/2.0.0</org.iids.aos.install.path>
-        <as.dir>${user.home}/Documents/projects/agentscape</as.dir>
+        <as.dir>${user.home}/programming/agentscape</as.dir>
         <!-- optional -->
     </properties>
     <repositories>
@@ -76,8 +76,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.3</version>
                 <configuration>
-                    <source>1.7</source>
-                    <target>1.7</target>
+                    <source>1.8</source>
+                    <target>1.8</target>
                 </configuration>
             </plugin>
             <!--
